app/requirements.txt
fastapi
uvicorn
openai
google-generativeai
requests
beautifulsoup4
python-multipart
pydantic
pdfplumber